Cleanup for PR into dev-branch
