--- src/remys_gift.rs
@@ -1,17 +1,0 @@
-// impl<Bs: Bits, M: Mode> Expr for Add<BitString<Bs, M>, _0, Eval>
-// {
-//     type Output = BitString<Bs, M>;
-// }
-// impl<Bs: Bits, M: Mode> Expr for Add<_0, BitString<Bs, M>, Eval>
-// {
-//     type Output = BitString<Bs, M>;
-// }
-
-// impl Expr for Add<BitString<Top1>, BitString<Top1>, Eval>
-// {
-//     type Output = BitString<More<Top1, _0>>;
-// }
-
-// impl Number for BitString<Bs, M> {
-//
-// }
